TIG-2064: Add capabilities for collection scanner and rolling collections. (#295)
New options: CollectionSkip (skip a number of collections when scanning),
CollectionSortOrder (sort the collection names forward or backward before
scanning), ScanDuration (hold a snapshot scan transaction open until a
duration has elapsed).
Modified option: Allow Database to be a single database name or a comma
separated list of databases to scan.
Rolling Collections now are named in a manner that gives a time based
ordering when sorted by name. This is needed by the large scale workload
so collections can be traversed "ahead" of when they are expiring.
Includes test for Collection Scanner.