PERF-2102 Update resharding key pattern used in ReshardCollection.yml.
Also lower the number of chunks. The new initial split calculated by the
reshardCollection command uses the number of existing chunks for the
collection being resharded as the default value for the numInitialChunks
parameter. {numInitialChunks: 60} leads to the an error about the new
shard key having a low cardinality because it would leave each chunk
with only 1 document.

Closes #450.