PERF-3485 Extend 'CrudActor' to support the aggregate command (#761)
PERF-3485 Extend 'CrudActor' to support the aggregate command

Aggregate, like find, returns a cursor. We often want to
measure the time taken to fully iterate the cursor,
including the time to process the initial request as well as
any subsequent getMore commands. Past workarounds obviated
by this change include using manipulating the batchSize to
ensure that the complete reply fits in a single batch, or
using explain with "executionStats" verbosity.