EVG-18409: Updates to get genny to work with new toolchain on m1
This involved upgrading cpp files to use the new version of catch that is
part of our toolchain. This directly depends on the catch cmake package
rather than using an intermediary library.

Additionally, there are some changes to handling triples and where
we install stuff as part of our setup scripts

This also rebuilds the poplar libs from grpc.