EVG-19133: Pre-warm connection for Loader.cpp.
Since connections on clients in libmongoc are opened lazily,
currently the connection time is being counted when we measure
time to load initial data. This caused BF-27679 when we
updated our infrastructure to have a valid certificate chain,
likely due to OpenSSL doing more work to process opening a TLS
connection. The recommended fix was to make sure our connections
are warm and open before we measure database operation, since
that steady state is what we care most about for user operations.

This change pings the database at the beginning of each phase to
make sure that we already have an active connection before we
time anything. This should bring timings down and reduce noise.