EVG-17874: Update Genny with --no-activate flag.
This allows us to create task generation tasks that don't
immediately execute all the generated variants.

It's a bit awkward to pass through a flag that is phrased in the
negative (as opposed to using an "activate" flag and variable),
but in this case we need to because an explicit activate flag
would have different behavior than the default of not specifying
the "activate" parameter. If the "activate" parameter is present
on a task, then that task is always activated, but if it is not
present then it defaults to if the version is activated or through
batchtime.