Clean up auto_tasks.
Get rid of CliOperation, because it was a mixed bag of random
data. Put build info in CurrentBuildInfo, and generally lift
reading the YAML expansion files out of as many places as possible.

This will make the code more flexible to being used outside of the
context of parsing a particular, currently running, build.