PERF-4478 Generator for an array of distinct values (#957)
PERF-4478 Generator for an array of distinct values

Whats Changed:
Add the `distinct` flag to ArrayGenerator. This will filter out duplicate generated values by using a set to check for duplicity. For simplicity, all generated values are stored in the set as `array::value` objects, so duplicity is determined by the hash function (defined in `Appendable::SetHasher`) and array value equality (defined in `Appendable::SetKeyEq`). There is a bound on the number of consecutive duplicates that are considered acceptable before giving up and erroring.

Also implements a refactor for the parsers maps to deduplicate maps and hedge against the risk of accidentally leaving a parser off of allParsers.

Patch testing results:
https://spruce.mongodb.com/version/64cc1a7357e85ac8087272bd/tasks