DEVPROD-1625 compile genny on al2023.0
Currently, cannot run genny on al2023.2 due to version of mongodb toolchain
being used by evergreen. The error is the following (on al2023.2, the available
SO verison is libffi.so.8):

```
ImportError: libffi.so.6: cannot open shared object file: No such file or
directory
```

The lastest mongodb toolchain has alrady been rebuilt on al2023.2 to fix this
issue. Hence, when evergreen adopts the latest toolchain (an indication could be
being unable to find libffi.so.8), update the `run_on` host to
`amazon2023.2-...`.