SERVER-85660: fix streams passthrough change stream source workload
The change stream source should only be consumed after all documents have been
written and durably persisted in the mongo collection, so the stream processor
should be started after all the documents have been added to the source
collection. In order to this we need to use `startAtOperationTime` in the
$source.config to tell the stream processor to consume the documents in the
change stream source from the beginning. It doesn't seem like there is a great
way to generate a BSON timestamp object that mongo IDL likes through genny YAML
or through the existing generators, so added `{^NowTimestamp: {}}` generator for
BSON timestamps, similar to BSON date generators. (NOTE: None of the BSON date
generators work here because the mongo IDL explicitly wants a BSON timestamp
object for `startAtOperationTime`).