PERF-5302: Properly process port number environment variables captured by Poplar
This fixes a potential bug with the Genny hook via Curator into Poplar.
[Poplar](https://github.com/mongodb/curator/blob/a1293258b30e04ea503e6b91b6b1628789e90093/operations/poplar.go#L21C34-L21C49) accepts the `POPLAR_RPC_HOST` and the
`POPLAR_RPC_PORT` environment variables and defaults to `localhost` and `2288`
if those environment variables are not set.

This change properly handles the `POPLAR_RPC_PORT` environment variable when setting
up Genny to connect to Poplar. Prior to this change, Genny would attempt to connect
to port `2288`, regardless of where the Poplar service was actually listening.