DEVPROD-6453 Fix MultikeyIndexes.yml indexes so that multi-planning is exercised as intended (#1188)
I noticed that all of the queries were using COLLSCAN plans. I think
this helps to explain the current results for this workload presented
here:
https://github.com/10gen/product-perf-experimentations/blob/master/investigations/PERF-5121-compare-multiplanners/full-results.ipynb.
Hopefully the results tell us something more meaningful about the three
multi-planning code paths after this fix!

I'm verifying that the workload runs correctly locally, but haven't done
any other testing of this change.

I will alert the perf build barons that change points (probably
improvements) are expected for this workload as a consequence of this
change!