PERF-5094 Update toolchain version for libmongocrypt Range V2 support (#1181)
**Jira Ticket:** PERF-5094

**Whats Changed:**  
Updates the dependency for libmongocrypt to a recent version for Range
V2 support. This is necessary for running the performance tests
described in the linked ticket. When the next version of libmongocrypt
releases, we plan to do another update to the toolchain version so that
we don't have a git-hash version in the toolchain; see
https://jira.mongodb.org/browse/PERF-5064

**Related PRs:**   
PR for toolchain change: https://github.com/10gen/vcpkg/pull/41