PERF-5241 Update AutoRun in workloads relevant to the QueryStats 6.0 backport (#1187)
### What's Changed
Updating the auto-run branches on tasks that we added query stats
auto-run configs for in PERF-5152 to include 6.0. Also, adding a query
stats variant to the MetricSecondaryIndexTimeseriesCollection.yml
auto-run because we'd like to get data from it on both backport
branches.

SERVER-87943 is adding the query stats variants that these will be run
under to v6.0.

### Patch Testing Results

https://spruce.mongodb.com/version/6617fa2e8ce8480007b1695c/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC