DEVPROD-5933 Validate team names and slack contact information for workloads. (#1195)
**Jira Ticket:** DEVPROD-5933 

### Whats Changed

For the workload ownership project we are going to require that all
workloads are owned by a valid team and have valid contact information
for support. We are using the Mothra repo as our source of truth for
this information. This PR introduces a `MothraService` that uses the
GitHub API to gather the yaml files representing teams at Mongo and
cross checks the owners and slack channels in each workload against this
data.

I created [this
spreadsheet](https://docs.google.com/spreadsheets/d/1e02apqAU77fC-cQ-HxvITKpa-6ch1U5ptE3SxdcbSYM/edit#gid=0)
to map existing workload owners to a team represented in Mothra and
gathered slack channels for support for each team. I updated all the
existing workload owners to their Mothra representation and added
contact information. The two exceptions to this rule are
`@mongodb/query` and `@mongodb/product-query` as these two need updated
to one of `Query Optimization`, `Query Integration`, or `Query
Execution`. There are 336 such instances that still need updated and I
will work with a query team member to fix these.


### Testing

The `./run-genny lint-yaml` command succeeds locally for me as I have a
GH token configured locally. The PR check is currently failing because I
don't yet have permissions to add a secret to the Evergreen project and
I'm waiting on IT to issue me an API key.


![image](https://github.com/mongodb/genny/assets/16945037/14edf1a3-8c4c-4afd-bacf-bc17a7b84402)



[Here](https://spruce.mongodb.com/task/genny_amazon2_t_lint_workloads_standalone_patch_67f55a09328efd294de9fa9fbdf5243bef1f27f4_6622f06acbb2230007d6d406_24_04_19_22_30_03/logs?execution=0)
is the new standalone task that I ran. It failed due to the access token
not being set yet.



###  Prerequisites

- Before merging we need [this
PR](https://github.com/10gen/mothra/pull/1003) in Mothra to be merged as
it includes the Slack support channels necessary for our validation.
- I have an open request with IT to get a Github token created to use
for the Genny evergreen project. This secret needs to be configured
before merging.

---------

Co-authored-by: Austin Hartschen <ahartschen@users.noreply.github.com>