PERF-5358 Improve CompoundIndexes.yml workload (#1214)
**Jira Ticket:** [PERF-5358](https://jira.mongodb.org/browse/PERF-5358)

### Whats Changed

There are three related changes:

1. Change the data so that there are 300,000 with approximately 100,000
associated with each of three tenants. Then change how the "selectivity"
value is calculated so that the query returns ~101 results. I did
confirm by looking at the mongod logs that beforehand the query was
returning 0 results and now it returns 87 (which is much closer to the
expected 101).
2. Change the predicate to have a conjunct for each indexed field. This
makes the test much more similar to "Simple.yml", and it's not clear to
me why it wasn't written this way in the first place.
3. Reduce the number of repetitions of each query from 1000 to 500. I
did this because adding the additional conjuncts made multi-planning run
for significantly longer, so I wanted to avoid making the workload run
for longer than necessary.

### Patch Testing Results

https://spruce.mongodb.com/version/663bb9ac889ffa0007942ecb/tasks