PERF-5375 delete NoGoodPlan.yml multi-planning workload (#1226)
**Jira Ticket:** [PERF-5375](https://jira.mongodb.org/browse/PERF-5375)

### Whats Changed

I've opted to delete the workload because it seems redundant with
NoSuchField.yml, which I'm improving in related PR
https://github.com/mongodb/genny/pull/1224. The scenarios are slightly
different because in NoSuchField.yml the extra predicate matches
nothing, whereas in this case the extra predicate does match some data
but cannot be answered efficiently. Still, the rationale for testing
both of these cases is more or less the same so they seem redundant.

I did consider applying the improvements originally described by this
ticket to NoSuchField.yml -- namely, the idea of increasing the data
size so that the "max works" value gets set based `0.3 *
collectionSize`. I ultimately decided against doing so because it makes
the workload take a lot longer to run. Just generating the data set and
building the indexes took a few minutes, and a single invocation of the
query with 64 indexes took over 2 minutes. It seems more important to me
to allow the workload to run a few minutes total rather than covering
the case where max works exceeds 10,000.

### Patch Testing Results

As far as I know, just the automatically executed genny self-tests
should be sufficient.