DEVPROD-7874: Update Genny to use the latest version of Curator (#1229)
**Jira Ticket:**
[DEVPROD-7874](https://jira.mongodb.org/browse/DEVPROD-7874)

### Whats Changed

Removed the special casing for Curator versions. Now, Genny always just
grabs the version of Curator pinned below. Also, removed the testing for
Ubuntu 1804 - it's EOL and as far as I can tell, we don't run any tests
on that os anymore. Please correct me if I'm wrong on that though.

### Patch Testing Results


[Genny](https://spruce.mongodb.com/version/6661cdaf7f45dc0007653b7e/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)

[Sys-perf

](https://spruce.mongodb.com/version/6661dc545f5be30007113ef2/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)

---------

Co-authored-by: Jeff Zambory <jeff.zambory@mongodb.com>