PERF-5355 Clean up query stats genny auto-run (#1228)
**Jira Ticket:** [PERF-5355](https://jira.mongodb.org/browse/PERF-5355)

### Whats Changed

Since the query stats backport is pretty much wrapped up, this PR will
cleanup the modifications made to auto run configurations in different
workloads to go back to the original states pre-project. The QueryStats
workload will be left as >=v7.0 since we'll be [getting
rid](https://jira.mongodb.org/browse/PERF-5456) of the `*-query-stats`
configurations in DSI.

### Patch Testing Results


[7.0](https://spruce.mongodb.com/version/665f6fcdd17bd600079ff5da/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)

[master](https://spruce.mongodb.com/version/6660a8085455de0007c0cbe6/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)

The patches are only running the QueryStats and QueryStatsQueryShapes
workloads, but the query stats variants are gone.