PERF-5628 Reduce runtime of TimeseriesEnum test (#1238)
**Jira Ticket:** [PERF-5628](https://jira.mongodb.org/browse/PERF-5628)

### Whats Changed

Changing the test to run on a smaller data set. This will help avoid
timeouts (some of the queries currently take 6+ minutes). At the same
time, this will not make the queries run so quickly that we don't get a
good signal on the query's performance.

### Patch Testing Results

Patch build running here:

https://spruce.mongodb.com/version/667c46abea070d00079381c4/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC