DEVPROD-11940 Add a flag to suppress metrics from the RunCommand actor (#1269)
**Jira Ticket:**
[DEVPROD-11940](https://jira.mongodb.org/browse/DEVPROD-11940)

### What's Changed

This PR adds a `ReportMetrics ` flag to the RunCommand actor that is
true by default. Additionally, I went through all uses of the RunCommand
actor and labeled anything that was clearly not a measurement as
`ReportMetrics: false`. This is usually indicated by Repeat: 1, and the
command creating indexes/dropping index/creating collections or setting
database parameters.

### Patch Testing Results
https://spruce.mongodb.com/version/670690abed45040007c733dc/