DEVPROD-18248 allow more time for curator startup (#1312)
The `t_compile_and_dry_run` CI task has been consistently failing on  
macOS Big Sur arm64. The root cause is that the Poplar process  
requires additional time to initialise on this specific platform.  

This PR increases the number of retry attempts for Poplar readiness to
`50`,
which extends the maximum wait time from `2` seconds to `10` seconds.
This
adjustment allows the task to succeed consistently.